requirements.txt
adabound==0.0.5
cnn-finetune==0.6.0
feather-format==0.4.0
japanize-matplotlib==1.0.5
lightgbm==2.3.0
logzero==1.5.0
numpy==1.17.2
opencv-python-headless==4.1.1.26
optuna==0.17.1
pandas==0.25.1
Pillow==6.2.0
protobuf==3.10.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboardX==1.9
torch==1.2.0
torchvision==0.4.0
tqdm==4.36.1
